feat(satellite): split api features in modules
