+++ src/libs/satellite/src/api/mod.rs
@@ -1,0 +1,7 @@
+pub mod cdn;
+pub mod config;
+pub mod controllers;
+pub mod db;
+pub mod http;
+pub mod rules;
+pub mod storage;
